refactor: move loop to file
